Revert "Hojjat | Mitchel / Add signup onboarding for CR  (#3509)"
This reverts commit 8e34b0ef4c796cc6eb9a8ca7fbb82d02106a517c.
