+++ packages/components/src/components/dialog/dialog.jsx
@@ -3,7 +3,6 @@ import classNames from 'classnames';
 import PropTypes from 'prop-types';
 import ReactDOM from 'react-dom';
 import { CSSTransition } from 'react-transition-group';
-import { isMobile } from '@deriv/shared';
 import Button from '../button/button.jsx';
 import Icon from '../icon/icon.jsx';
 import Text from '../text';
@@ -18,7 +17,6 @@ const Dialog = ({
     is_visible,
     onCancel,
     onConfirm,
-    header_icon,
     ...other_props
 }) => {
     const wrapper_ref = React.useRef();
@@ -80,51 +78,26 @@ const Dialog = ({
             <div
                 className={classNames('dc-dialog__wrapper', className, {
                     'dc-dialog__wrapper--has-portal': !!portal_element_id,
-                    'dc-dialog__wrapper--has-icon': header_icon,
                 })}
             >
                 <div
                     className={classNames('dc-dialog__dialog', {
                         'dc-dialog__dialog--has-margin': !is_mobile_full_width,
-                        'dc-dialog__dialog--no-padding': header_icon,
                     })}
                     ref={wrapper_ref}
                 >
-                    {header_icon && title ? (
-                        <div className='dc-dialog__header-wrapper__centered'>
-                            <Icon
-                                className='dc-dialog__header-verification-icon'
-                                icon={header_icon}
-                                size={isMobile() ? 30 : 35}
-                            />
-                            <Text
-                                as='h1'
-                                align='center'
-                                color='prominent'
-                                weight='bold'
-                                className='dc-dialog__header'
-                                size={isMobile() ? 'xs' : 'sm'}
-                            >
+                    <div className='dc-dialog__header-wrapper'>
+                        {!!title && (
+                            <Text as='h1' color='prominent' weight='bold' className='dc-dialog__header'>
                                 {title}
                             </Text>
-                        </div>
-                    ) : (
-                        <div className='dc-dialog__header-wrapper'>
-                            {!!title && (
-                                <Text as='h1' color='prominent' weight='bold' className='dc-dialog__header'>
-                                    {title}
-                                </Text>
-                            )}
-                            {has_close_icon && (
-                                <div
-                                    onClick={onCancel ? handleCancel : handleConfirm}
-                                    className='dc-dialog__header--close'
-                                >
-                                    <Icon icon='IcCross' />
-                                </div>
-                            )}
-                        </div>
-                    )}
+                        )}
+                        {has_close_icon && (
+                            <div onClick={onCancel ? handleCancel : handleConfirm} className='dc-dialog__header--close'>
+                                <Icon icon='IcCross' />
+                            </div>
+                        )}
+                    </div>
                     {is_text ? (
                         <Text as='p' size='xs' styles={{ lineHeight: '1.43' }} className={content_classes}>
                             {children}
@@ -169,7 +142,6 @@ const Dialog = ({
 Dialog.defaultProps = {
     is_closed_on_cancel: true,
     is_closed_on_confirm: true,
-    header_icon: false,
 };
 
 Dialog.propTypes = {
@@ -178,7 +150,6 @@ Dialog.propTypes = {
     disableApp: PropTypes.func,
     enableApp: PropTypes.func,
     has_close_icon: PropTypes.bool,
-    header_icon: PropTypes.string,
     is_closed_on_cancel: PropTypes.bool,
     is_closed_on_confirm: PropTypes.bool,
     is_content_centered: PropTypes.bool,
