+++ packages/components/src/components/icon/icons.js
@@ -140,7 +140,6 @@ import './common/ic-alert-danger.svg';
 import './common/ic-alert-info.svg';
 import './common/ic-alert-success.svg';
 import './common/ic-alert-warning.svg';
-import './common/ic-android.svg';
 import './common/ic-archive.svg';
 import './common/ic-arrow-down.svg';
 import './common/ic-arrow-left-bold.svg';
@@ -157,7 +156,6 @@ import './common/ic-bank.svg';
 import './common/ic-bell.svg';
 import './common/ic-black-warning.svg';
 import './common/ic-box.svg';
-import './common/ic-browser.svg';
 import './common/ic-calendar-datefrom.svg';
 import './common/ic-calendar-dateto.svg';
 import './common/ic-calendar-forward-today.svg';
@@ -241,18 +239,15 @@ import './common/ic-installation-macos.svg';
 import './common/ic-installation-web-demo.svg';
 import './common/ic-installation-web-real.svg';
 import './common/ic-installation-windows.svg';
-import './common/ic-ios.svg';
 import './common/ic-language.svg';
 import './common/ic-less-than-eight.svg';
 import './common/ic-linux.svg';
-import './common/ic-linux2.svg';
 import './common/ic-live-chat.svg';
 import './common/ic-lock.svg';
 import './common/ic-logout.svg';
 import './common/ic-lookbacks.svg';
 import './common/ic-loss.svg';
 import './common/ic-macos.svg';
-import './common/ic-macos2.svg';
 import './common/ic-message-delivered.svg';
 import './common/ic-message-errored.svg';
 import './common/ic-message-pending.svg';
@@ -349,10 +344,8 @@ import './common/ic-user-outline.svg';
 import './common/ic-user.svg';
 import './common/ic-utility.svg';
 import './common/ic-verification.svg';
-import './common/ic-verify.svg';
 import './common/ic-web-terminal.svg';
 import './common/ic-website.svg';
-import './common/ic-windows-gray.svg';
 import './common/ic-windows.svg';
 import './common/ic-wip.svg';
 import './common/ic-zoom-in.svg';
@@ -430,8 +423,6 @@ import './flag/ic-flag-uk.svg';
 import './flag/ic-flag-vi.svg';
 import './flag/ic-flag-zh-cn.svg';
 import './flag/ic-flag-zh-tw.svg';
-import './mt5/ic-mt5-commodities.svg';
-import './mt5/ic-mt5-cryptos.svg';
 import './mt5/ic-mt5-device-desktop.svg';
 import './mt5/ic-mt5-device-laptop.svg';
 import './mt5/ic-mt5-device-phone.svg';
@@ -441,22 +432,18 @@ import './mt5/ic-mt5-financial-platform.svg';
 import './mt5/ic-mt5-financial-stp-platform.svg';
 import './mt5/ic-mt5-financial-stp.svg';
 import './mt5/ic-mt5-financial.svg';
-import './mt5/ic-mt5-forex.svg';
 import './mt5/ic-mt5-high-leverage.svg';
-import './mt5/ic-mt5-indices.svg';
 import './mt5/ic-mt5-liquid-risk.svg';
 import './mt5/ic-mt5-margin-trading.svg';
 import './mt5/ic-mt5-one-password.svg';
 import './mt5/ic-mt5-open-markets.svg';
 import './mt5/ic-mt5-password-updated.svg';
 import './mt5/ic-mt5-responsive.svg';
-import './mt5/ic-mt5-stocks.svg';
 import './mt5/ic-mt5-support.svg';
 import './mt5/ic-mt5-synthetic-dashboard.svg';
 import './mt5/ic-mt5-synthetic-indices.svg';
 import './mt5/ic-mt5-synthetic-platform.svg';
 import './mt5/ic-mt5-synthetic.svg';
-import './mt5/ic-mt5-synthetics.svg';
 import './mt5/ic-mt5-trade-types.svg';
 import './option/ic-option-call-put-reset.svg';
 import './option/ic-option-call-put-spread.svg';
