+++ packages/components/src/components/table/table.jsx
@@ -2,7 +2,6 @@ import React from 'react';
 import classNames from 'classnames';
 import Head from './table-head.jsx';
 import Header from './table-header.jsx';
-import Footer from './table-footer.jsx';
 import Body from './table-body.jsx';
 import Row from './table-row.jsx';
 import Cell from './table-cell.jsx';
@@ -36,6 +35,6 @@ Table.Header = Header;
 Table.Body = Body;
 Table.Row = Row;
 Table.Cell = Cell;
-Table.Footer = Footer;
+// TODO add footer
 
 export default Table;
