+++ packages/components/src/components/table/table.scss
@@ -62,8 +62,4 @@
             display: unset;
         }
     }
-    &__footer {
-        border-bottom: none;
-        height: 5.2rem;
-    }
 }
