+++ packages/components/stories/icon/icons.js
@@ -146,7 +146,6 @@ export const icons = {
         'IcAlertInfo',
         'IcAlertSuccess',
         'IcAlertWarning',
-        'IcAndroid',
         'IcArchive',
         'IcArrowDown',
         'IcArrowLeftBold',
@@ -163,7 +162,6 @@ export const icons = {
         'IcBell',
         'IcBlackWarning',
         'IcBox',
-        'IcBrowser',
         'IcCalendarDatefrom',
         'IcCalendarDateto',
         'IcCalendarForwardToday',
@@ -247,18 +245,15 @@ export const icons = {
         'IcInstallationWebDemo',
         'IcInstallationWebReal',
         'IcInstallationWindows',
-        'IcIos',
         'IcLanguage',
         'IcLessThanEight',
         'IcLinux',
-        'IcLinux2',
         'IcLiveChat',
         'IcLock',
         'IcLogout',
         'IcLookbacks',
         'IcLoss',
         'IcMacos',
-        'IcMacos2',
         'IcMessageDelivered',
         'IcMessageErrored',
         'IcMessagePending',
@@ -355,10 +350,8 @@ export const icons = {
         'IcUser',
         'IcUtility',
         'IcVerification',
-        'IcVerify',
         'IcWebTerminal',
         'IcWebsite',
-        'IcWindowsGray',
         'IcWindows',
         'IcWip',
         'IcZoomIn',
@@ -448,8 +441,6 @@ export const icons = {
         'IcFlagZhTw',
     ],
     mt5: [
-        'IcMt5Commodities',
-        'IcMt5Cryptos',
         'IcMt5DeviceDesktop',
         'IcMt5DeviceLaptop',
         'IcMt5DevicePhone',
@@ -459,22 +450,18 @@ export const icons = {
         'IcMt5FinancialStpPlatform',
         'IcMt5FinancialStp',
         'IcMt5Financial',
-        'IcMt5Forex',
         'IcMt5HighLeverage',
-        'IcMt5Indices',
         'IcMt5LiquidRisk',
         'IcMt5MarginTrading',
         'IcMt5OnePassword',
         'IcMt5OpenMarkets',
         'IcMt5PasswordUpdated',
         'IcMt5Responsive',
-        'IcMt5Stocks',
         'IcMt5Support',
         'IcMt5SyntheticDashboard',
         'IcMt5SyntheticIndices',
         'IcMt5SyntheticPlatform',
         'IcMt5Synthetic',
-        'IcMt5Synthetics',
         'IcMt5TradeTypes',
     ],
     option: [
