+++ packages/core/src/App/Containers/AccountSignupModal/index.js
@@ -1,0 +1,1 @@
+export default from './account-signup-modal.jsx';
