+++ packages/core/src/App/Containers/Layout/app-contents.jsx
@@ -1,10 +1,10 @@
 import classNames from 'classnames';
 import PropTypes from 'prop-types';
 import React from 'react';
-import { useLocation } from 'react-router';
+import { withRouter } from 'react-router';
 import WS from 'Services/ws-methods';
 import { DesktopWrapper, MobileWrapper, ThemedScrollbars } from '@deriv/components';
-import { CookieStorage, isMobile, TRACKING_STATUS_KEY, PlatformContext, routes } from '@deriv/shared';
+import { CookieStorage, isMobile, TRACKING_STATUS_KEY, PlatformContext } from '@deriv/shared';
 import { connect } from 'Stores/connect';
 import CookieBanner from '../../Components/Elements/CookieBanner/cookie-banner.jsx';
 
@@ -29,10 +29,8 @@ const AppContents = ({
     const [show_cookie_banner, setShowCookieBanner] = React.useState(false);
     const [is_gtm_tracking, setIsGtmTracking] = React.useState(false);
     const { is_dashboard } = React.useContext(PlatformContext);
-    const { pathname } = useLocation();
 
     const tracking_status = tracking_status_cookie.get(TRACKING_STATUS_KEY);
-    const is_onboarding = pathname === routes.onboarding;
 
     React.useEffect(() => {
         const allow_tracking = !is_eu_country || tracking_status === 'accepted';
@@ -88,15 +86,14 @@ const AppContents = ({
                 'app-contents--is-disabled': is_app_disabled,
                 'app-contents--is-mobile': isMobile(),
                 'app-contents--is-route-modal': is_route_modal_on,
-                'app-contents--is-scrollable': is_cfd_page || is_cashier_visible || is_onboarding,
+                'app-contents--is-scrollable': is_cfd_page || is_cashier_visible,
                 'app-contents--is-dashboard': is_dashboard,
-                'app-contents--is-onboarding': is_onboarding,
             })}
         >
             <MobileWrapper>{children}</MobileWrapper>
             <DesktopWrapper>
                 {/* Calculate height of user screen and offset height of header and footer */}
-                <ThemedScrollbars height={is_onboarding ? 'calc(100vh - 72px)' : 'calc(100vh - 84px)'} has_horizontal>
+                <ThemedScrollbars height='calc(100vh - 84px)' has_horizontal>
                     {children}
                 </ThemedScrollbars>
             </DesktopWrapper>
@@ -122,19 +119,21 @@ AppContents.propTypes = {
     is_route_modal_on: PropTypes.bool,
 };
 
-export default connect(({ client, gtm, rudderstack, ui }) => ({
-    is_eu_country: client.is_eu_country,
-    is_eu: client.is_eu,
-    is_logged_in: client.is_logged_in,
-    is_logging_in: client.is_logging_in,
-    pushDataLayer: gtm.pushDataLayer,
-    identifyEvent: rudderstack.identifyEvent,
-    pageView: rudderstack.pageView,
-    is_app_disabled: ui.is_app_disabled,
-    is_cashier_visible: ui.is_cashier_visible,
-    is_dark_mode: ui.is_dark_mode_on,
-    is_cfd_page: ui.is_cfd_page,
-    is_positions_drawer_on: ui.is_positions_drawer_on,
-    is_route_modal_on: ui.is_route_modal_on,
-    notifyAppInstall: ui.notifyAppInstall,
-}))(AppContents);
+export default withRouter(
+    connect(({ client, gtm, rudderstack, ui }) => ({
+        is_eu_country: client.is_eu_country,
+        is_eu: client.is_eu,
+        is_logged_in: client.is_logged_in,
+        is_logging_in: client.is_logging_in,
+        pushDataLayer: gtm.pushDataLayer,
+        identifyEvent: rudderstack.identifyEvent,
+        pageView: rudderstack.pageView,
+        is_app_disabled: ui.is_app_disabled,
+        is_cashier_visible: ui.is_cashier_visible,
+        is_dark_mode: ui.is_dark_mode_on,
+        is_cfd_page: ui.is_cfd_page,
+        is_positions_drawer_on: ui.is_positions_drawer_on,
+        is_route_modal_on: ui.is_route_modal_on,
+        notifyAppInstall: ui.notifyAppInstall,
+    }))(AppContents)
+);
