--- packages/core/src/App/Containers/Onboarding/index.js
@@ -1,1 +1,0 @@
-export Onboarding from './onboarding.jsx';
