+++ packages/core/src/App/Containers/Redirect/redirect.jsx
@@ -1,6 +1,7 @@
 import PropTypes from 'prop-types';
+import React from 'react';
 import { withRouter } from 'react-router-dom';
-import { loginUrl, routes } from '@deriv/shared';
+import { loginUrl, routes, PlatformContext } from '@deriv/shared';
 import { getLanguage } from '@deriv/translations';
 import { connect } from 'Stores/connect';
 import { WS } from 'Services';
@@ -12,27 +13,27 @@ const Redirect = ({
     hasAnyRealAccount,
     openRealAccountSignup,
     setResetTradingPasswordModalOpen,
+    toggleAccountSignupModal,
     toggleResetPasswordModal,
-    logout,
-    is_logged_in,
 }) => {
     const url_query_string = window.location.search;
     const url_params = new URLSearchParams(url_query_string);
     let redirected_to_route = false;
+    const { is_dashboard } = React.useContext(PlatformContext);
 
     setVerificationCode(url_params.get('code'), url_params.get('action'));
 
     switch (url_params.get('action')) {
         case 'signup': {
-            history.push({
-                pathname: routes.onboarding,
-                search: url_query_string,
-            });
-            redirected_to_route = true;
-            sessionStorage.removeItem('redirect_url');
-            if (is_logged_in) {
-                logout();
+            if (is_dashboard) {
+                history.push({
+                    pathname: routes.dashboard,
+                    search: url_query_string,
+                });
+                redirected_to_route = true;
             }
+            sessionStorage.removeItem('redirect_url');
+            toggleAccountSignupModal(true);
             break;
         }
         case 'reset_password': {
@@ -112,8 +113,7 @@ export default withRouter(
         hasAnyRealAccount: client.hasAnyRealAccount,
         openRealAccountSignup: ui.openRealAccountSignup,
         setResetTradingPasswordModalOpen: ui.setResetTradingPasswordModalOpen,
+        toggleAccountSignupModal: ui.toggleAccountSignupModal,
         toggleResetPasswordModal: ui.toggleResetPasswordModal,
-        is_logged_in: client.is_logged_in,
-        logout: client.logout,
     }))(Redirect)
 );
