+++ packages/core/src/sass/app/modules/welcome.scss
@@ -9,12 +9,6 @@
 
 .welcome-cr {
     background-color: var(--general-section-4);
-    padding: 2.4rem 1.6rem 0;
-    height: 100vh;
-
-    @include desktop {
-        padding: 10.2rem 0 0;
-    }
 }
 
 .welcome {
@@ -22,7 +16,6 @@
 
     &__header {
         margin: 2.4rem;
-        margin-top: 0;
     }
 
     &__main {
@@ -81,12 +74,6 @@
             align-items: center;
             flex-basis: 32px;
             display: flex;
-
-            @include mobile {
-                position: absolute;
-                top: 3.4rem;
-                right: 0.5rem;
-            }
         }
 
         &__body {
@@ -105,12 +92,6 @@
             display: flex;
             flex-direction: column;
             justify-content: center;
-
-            &-description {
-                @include mobile {
-                    margin-right: 1.7rem;
-                }
-            }
         }
 
         &__options {
@@ -119,7 +100,7 @@
 
         &__option {
             padding: 0 0.2rem;
-            border-right: 1px solid var(--border-hover);
+            border-right: 1px solid var(--general-section-1);
             line-break: anywhere;
 
             &:first-of-type {
