+++ packages/shared/src/styles/constants.scss
@@ -156,14 +156,12 @@ $MAX_CONTAINER_WIDTH: 1440px;
 
 $POSITIONS_DRAWER_WIDTH: 240px;
 $POSITIONS_DRAWER_MARGIN: 8px;
-$HEADER_HEIGHT_SIGNUP: 72px;
 
 $HEADER_HEIGHT: 48px;
 $FOOTER_HEIGHT: 36px;
 
 $MOBILE_HEADER_HEIGHT: 40px;
 $MOBILE_HEADER_HEIGHT_DASHBOARD: 48px;
-$MOBILE_HEADER_HEIGHT_SIGNUP: 48px;
 $MOBILE_WRAPPER_HEADER_HEIGHT: 40px;
 $MOBILE_WRAPPER_FOOTER_HEIGHT: 70px;
