+++ packages/shared/src/utils/routes/routes.js
@@ -33,7 +33,6 @@ export const routes = {
     reports: '/reports',
     root: '/',
     redirect: '/redirect',
-    onboarding: '/onboarding',
     settings: '/settings',
     statement: '/reports/statement',
     token: '/settings/token',
